backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
openai==1.68.2
httpx==0.27.2
python-dotenv==1.0.0
